requirements.txt
djangorestframework~=3.14.0
Django~=4.2.3
numpy~=1.24.1
pandas~=1.5.3
matplotlib~=3.7.0
reportlab~=4.0.8
requests~=2.31.0
lxml~=4.9.4
PyMySQL~=1.1.0